feat udistrital/arka_cliente#1098: Seguridad
- Refactor modelos.
