+++ helpers/asientoContable/GenerarMovimientosContables.go
@@ -8,7 +8,7 @@ import (
 )
 
 // GenerarMovimientosContables Genera los movimientos contables para una serie de cuentas y valores
-func GenerarMovimientosContables(totales map[int]float64, detalleCuentas map[string]models.CuentaContable, cuentasSubgrupo map[int]models.CuentaSubgrupo,
+func GenerarMovimientosContables(totales map[int]float64, detalleCuentas map[string]models.CuentaContable, cuentasSubgrupo map[int]models.CuentasSubgrupo,
 	parDebito, parCredito, terceroIdCr, terceroIdDb int, descripcion string, ajuste bool, movimientos *[]*models.MovimientoTransaccion) (outputError map[string]interface{}) {
 
 	funcion := "GenerarMovimientosContables"
@@ -31,7 +31,7 @@ func GenerarMovimientosContables(totales map[int]float64, detalleCuentas map[str
 }
 
 // ConstruirMovimientosContables Genera los movimientos contables para una serie de cuentas y valores
-func ConstruirMovimientosContables(totales map[int]float64, detalleCuentas map[string]models.CuentaContable, cuentasSubgrupo map[int]models.CuentaSubgrupo,
+func ConstruirMovimientosContables(totales map[int]float64, detalleCuentas map[string]models.CuentaContable, cuentasSubgrupo map[int]models.CuentasSubgrupo,
 	terceroIdCr, terceroIdDb int, descripcion string, ajuste bool, movimientos *[]*models.MovimientoTransaccion) (
 	err string, outputError map[string]interface{}) {
