+++ helpers/asientoContable/asientoContable.helper.go
@@ -47,7 +47,7 @@ func AsientoContable(totales map[int]float64, comprobante, tipomvto, descripcion
 		transaccion          models.TransaccionMovimientos
 		parametroTipoDebito  int
 		parametroTipoCredito int
-		cuentasSubgrupo      []*models.CuentaSubgrupo
+		cuentasSubgrupo      []*models.CuentasSubgrupo
 		comprobanteID        string
 	)
 
@@ -200,7 +200,7 @@ func fillDetalle(cuentas map[string]*models.CuentaContable, transaccion *models.
 }
 
 // findIdInArray Retorna la posicion en que se encuentra el id específicado
-func FindInArray(cuentasSg []*models.CuentaSubgrupo, subgrupoId int) (i int) {
+func FindInArray(cuentasSg []*models.CuentasSubgrupo, subgrupoId int) (i int) {
 	for i, cuentaSg := range cuentasSg {
 		if int(cuentaSg.SubgrupoId.Id) == subgrupoId {
 			return i
