+++ helpers/crud/catalogoElementos/catalogoElementos.go
@@ -13,7 +13,7 @@ import (
 var basePath = "http://" + beego.AppConfig.String("catalogoElementosService")
 
 // GetAllCuentasSubgrupo query controlador cuentas_subgrupo del api catalogo_elementos_crud
-func GetAllCuentasSubgrupo(query string) (elementos []*models.CuentaSubgrupo, outputError map[string]interface{}) {
+func GetAllCuentasSubgrupo(query string) (elementos []*models.CuentasSubgrupo, outputError map[string]interface{}) {
 
 	funcion := "GetAllCuentasSubgrupo"
 	defer errorctrl.ErrorControlFunction(funcion+" - Unhandled Error!", "500")
