--- models/Dependencia.go
@@ -1,8 +1,0 @@
-package models
-
-type Dependencia struct {
-	Id          				int
-	Nombre     					string
-	TelefonoDependencia        	string
-	CorreoElectronico 			string
-}
