+++ models/acta_recibido.go
@@ -55,3 +55,82 @@ type Elemento_campo struct {
 	FechaCreacion     time.Time
 	FechaModificacion time.Time
 }
+
+type TransaccionActaRecibido struct {
+	ActaRecibido *ActaRecibido
+	UltimoEstado *HistoricoActa
+	SoportesActa *[]SoporteActa
+	Elementos    []*Elemento
+}
+
+type SoporteActa struct {
+	Id                int
+	Consecutivo       string
+	DocumentoId       int
+	FechaSoporte      time.Time
+	ActaRecibidoId    *ActaRecibido
+	Activo            bool
+	FechaCreacion     time.Time
+	FechaModificacion time.Time
+}
+
+type HistoricoActa struct {
+	Id                int
+	ProveedorId       int
+	UbicacionId       int
+	RevisorId         int
+	PersonaAsignadaId int
+	Observaciones     string
+	FechaVistoBueno   time.Time
+	ActaRecibidoId    *ActaRecibido
+	EstadoActaId      *EstadoActa
+	Activo            bool
+	FechaCreacion     time.Time
+	FechaModificacion time.Time
+}
+
+type EstadoActa struct {
+	Id                int
+	Nombre            string
+	Descripcion       string
+	CodigoAbreviacion string
+	Activo            bool
+	NumeroOrden       float64
+	FechaCreacion     time.Time
+	FechaModificacion time.Time
+}
+
+type EstadoElemento struct {
+	Id                int
+	Nombre            string
+	Descripcion       string
+	CodigoAbreviacion string
+	Activo            bool
+	NumeroOrden       float64
+	FechaCreacion     time.Time
+	FechaModificacion time.Time
+}
+
+type Elemento struct {
+	Id                 int
+	Nombre             string
+	Cantidad           int
+	Marca              string
+	Serie              string
+	UnidadMedida       int
+	ValorUnitario      float64
+	Subtotal           float64
+	Descuento          float64
+	ValorTotal         float64
+	PorcentajeIvaId    int
+	ValorIva           float64
+	ValorFinal         float64
+	SubgrupoCatalogoId int
+	TipoBienId         int
+	EstadoElementoId   *EstadoElemento
+	ActaRecibidoId     *ActaRecibido
+	Placa              string
+	Activo             bool
+	FechaCreacion      time.Time
+	FechaModificacion  time.Time
+}
