--- models/asignacion_espacio_fisico_dependencia.go
@@ -1,13 +1,0 @@
-package models
-
-import "time"
-
-type AsignacionEspacioFisicoDependencia struct {
-	Id               int
-	EspacioFisicoId  *EspacioFisico
-	DependenciaId    *Dependencia
-	Activo           bool
-	FechaInicio      time.Time
-	FechaFin         time.Time
-	DocumentoSoporte int
-}
