--- models/estado_acta.go
@@ -1,14 +1,0 @@
-package models
-
-import "time"
-
-type EstadoActa struct {
-	Id                int
-	Nombre            string
-	Descripcion       string
-	CodigoAbreviacion string
-	Activo            bool
-	NumeroOrden       float64
-	FechaCreacion     time.Time
-	FechaModificacion time.Time
-}
