--- models/nivel_clasificacion.go
@@ -1,11 +1,0 @@
-package models
-
-type NivelClasificacion struct {
-	Id                int
-	Nombre            string
-	Longitud          int
-	Descripcion       string
-	CodigoAbreviacion string
-	Activo            bool
-	NumeroOrden       int
-}
