+++ models/oikos_api.go
@@ -1,18 +1,29 @@
 package models
 
+import "time"
+
+type AsignacionEspacioFisicoDependencia struct {
+	Id               int
+	EspacioFisicoId  *EspacioFisico
+	DependenciaId    *Dependencia
+	Activo           bool
+	FechaInicio      time.Time
+	FechaFin         time.Time
+	DocumentoSoporte int
+}
+
 type EspacioFisico struct {
 	Id                  int
 	Nombre              string
 	CodigoAbreviacion   string
 	Activo              bool
-	TipoEspacio         *TipoEspacio
 	Descripcion         string
 	TipoTerrenoId       int
 	TipoEdificacionId   int
-	TipoEspacioFisicoId *TipoEspacioFisicoV2
+	TipoEspacioFisicoId *TipoEspacioFisico
 }
 
-type TipoEspacioFisicoV2 struct {
+type TipoEspacioFisico struct {
 	Id                int
 	Nombre            string
 	Descripcion       string
@@ -24,4 +35,11 @@ type DetalleSedeDependencia struct {
 	Sede        *EspacioFisico
 	Dependencia *Dependencia
 	Ubicacion   *AsignacionEspacioFisicoDependencia
+}
+
+type Dependencia struct {
+	Id                  int
+	Nombre              string
+	TelefonoDependencia string
+	CorreoElectronico   string
 }
