--- models/soporte_acta.go
@@ -1,14 +1,0 @@
-package models
-
-import "time"
-
-type SoporteActa struct {
-	Id                int
-	Consecutivo       string
-	DocumentoId       int
-	FechaSoporte      time.Time
-	ActaRecibidoId    *ActaRecibido
-	Activo            bool
-	FechaCreacion     time.Time
-	FechaModificacion time.Time
-}
