--- models/transaccion_acta_recibido.go
@@ -1,8 +1,0 @@
-package models
-
-type TransaccionActaRecibido struct {
-	ActaRecibido *ActaRecibido
-	UltimoEstado *HistoricoActa
-	SoportesActa *[]SoporteActa
-	Elementos    []*Elemento
-}
